Fixed polymorphic tests
