+++ Akka.Serialization.Testkit/PolymorphismTests.cs
@@ -128,27 +128,6 @@ namespace Akka.Serialization.Testkit
             AssertEqual(actual);
         }
 
-        [Fact]
-        public virtual void Can_Serialize_internal_class()
-        {
-            var internalClass = new PolymorphismMessages.InternalClass(55);
-            AssertEqual(internalClass);
-        }
-
-        [Fact]
-        public virtual void Can_Serialize_a_class_with_internal_constructor()
-        {
-            var classWithInternalConstructor = new PolymorphismMessages.ClassWithInternalConstructor(55);
-            AssertEqual(classWithInternalConstructor);
-        }
-
-        [Fact]
-        public virtual void Can_Serialize_a_class_with_private_constructor()
-        {
-            var privateConstructor = PolymorphismMessages.ClassWithPrivateConstructor.Instance;
-            AssertEqual(privateConstructor);
-        }
-
         protected T AssertAndReturn<T>(T message)
         {
             var serializer = Sys.Serialization.FindSerializerFor(message);
@@ -219,57 +198,6 @@ namespace Akka.Serialization.Testkit
                     return ((Name != null ? Name.GetHashCode() : 0) * 397) ^ EqualityComparer<T>.Default.GetHashCode(Data);
                 }
             }
-        }
-
-        internal class InternalClass
-        {
-            public InternalClass(int number)
-            {
-                Number = number;
-            }
-
-            public int Number { get; }
-
-            protected bool Equals(InternalClass other) => Number == other.Number;
-
-            public override bool Equals(object obj)
-            {
-                if (ReferenceEquals(null, obj)) return false;
-                if (ReferenceEquals(this, obj)) return true;
-                if (obj.GetType() != this.GetType()) return false;
-                return Equals((InternalClass)obj);
-            }
-
-            public override int GetHashCode() => Number;
-        }
-
-        public class ClassWithPrivateConstructor
-        {
-            public static ClassWithPrivateConstructor Instance { get; } = new ClassWithPrivateConstructor();
-            private ClassWithPrivateConstructor() { }
-            public override bool Equals(object obj) => obj is ClassWithPrivateConstructor;
-        }
-
-        public class ClassWithInternalConstructor
-        {
-            internal ClassWithInternalConstructor(int number)
-            {
-                Number = number;
-            }
-
-            public int Number { get; }
-
-            protected bool Equals(ClassWithInternalConstructor other) => Number == other.Number;
-
-            public override bool Equals(object obj)
-            {
-                if (ReferenceEquals(null, obj)) return false;
-                if (ReferenceEquals(this, obj)) return true;
-                if (obj.GetType() != this.GetType()) return false;
-                return Equals((ClassWithInternalConstructor)obj);
-            }
-
-            public override int GetHashCode() => Number;
         }
     }
 }
